Update gorgias_switch.js
If a previous chat on Gorgias has been inactive for more than 1 hour, then Certainly is loaded AND the previous chat access on Gorgias is cleared.
This way, a new handover to Gorgias Chat will not result in the visitor resuming a previous conversation.
The issue with resuming a previous conversations is that previous Certainly transcripts would be shown, which would be bad UX